pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { 'verifit' = 'src/verifit' }
packages = ['verifit']

[tool.check-manifest]
ignore = [".pytest_cache/*", "node_modules/**/*"]

[project]
name = "verifit"
version = "1.0.4"
description = "Verify It: Automatic Testing helper tools & sample tests"
readme = "Readme.md"
authors = [{ name = "Sorel Mitra", email = "sorelmitra@yahoo.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["automatic testing", "acceptance testing", "bdd", "Gherkin", "black box testing"]
dependencies = [
    "PyJWT",
    "pytest",
    "pytest-bdd",
    "python-dotenv",
    "python_graphql_client",
    "requests",
    "websockets",
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = []

[project.urls]
Homepage = "https://github.com/sorelmitra/verifit"

